requirements.txt
wheel
flask
uwsgi
requests
hmac
GitPython